# www/app/Http/Requests/StoreTestGroupRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Contracts\Validation\Validator;
use Illuminate\Foundation\Http\FormRequest;
use Illuminate\Http\Exceptions\HttpResponseException;

/**
 * @OA\Schema(
 *      title="Store Test request",
 *      description="Store Test request body data",
 *      type="object",
 *      required={"group_name"}
 * )
 */
class StoreTestGroupRequest extends FormRequest
{
    /**
     * @OA\Property(
     *      title="Test name",
     *      description="Test name"
     * )
     *
     * @var string
     */
    public string $group_name;

    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, mixed>
     */
    public function rules()
    {
        return [
            'group_name' => 'required'
        ];
    }

    protected function failedValidation(Validator $validator)
    {
        throw new HttpResponseException(response()->json([
            'status' => false,
            'errors' => $validator->errors()
        ]));
    }
}