# www/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fiat Lux api documentation",
        "description": "Fiat Lux School OpenApi documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/test-groups": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Returns a list of tests",
                "description": "Display a listing of the resource.",
                "operationId": "6963f650c26fe67b201e09a52615a3fa",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TestGroup"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Unauthenticated",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tests"
                ],
                "summary": "Saves test name",
                "description": "Update the specified resource in storage.",
                "operationId": "e0e7dd5c9fc90259155f1a0d208cee66",
                "parameters": [
                    {
                        "name": "test_group",
                        "in": "path",
                        "description": "Test id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTestGroupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TestGroup"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tests"
                ],
                "summary": "Saves test name",
                "description": "Store a newly created resource in storage.",
                "operationId": "3a0d22c979e5da8865cd06787757b797",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTestGroupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TestGroup"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreTestGroupRequest": {
                "title": "Store Test request",
                "description": "Store Test request body data",
                "required": [
                    "group_name"
                ],
                "properties": {
                    "group_name": {
                        "title": "Test name",
                        "description": "Test name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Option": {
                "title": "Option Response model",
                "description": "Option Response model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Current test id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "text": {
                        "title": "Option text",
                        "description": "Option text",
                        "type": "string",
                        "format": "string"
                    },
                    "is_correct": {
                        "title": "Is correct option flag",
                        "description": "MArk if current option is correct for the question",
                        "type": "boolean",
                        "format": "boolean"
                    },
                    "test_id": {
                        "title": "Test id",
                        "description": "Test id for the current option",
                        "type": "integer",
                        "format": "int64"
                    },
                    "test": {
                        "$ref": "#/components/schemas/Test"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Option"
                }
            },
            "Test": {
                "title": "Test Response model",
                "description": "Test Response model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Current test id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "question": {
                        "title": "Test question",
                        "description": "Test question",
                        "type": "string",
                        "format": "string"
                    },
                    "test_group_id": {
                        "title": "Test group id",
                        "description": "Test group id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "test_group": {
                        "$ref": "#/components/schemas/TestGroup"
                    },
                    "options": {
                        "title": "List of options",
                        "description": "List of options for the current test",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Option"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Test"
                }
            },
            "TestGroup": {
                "title": "Test Group Response model",
                "description": "Test Group Response model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Current test group id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "group_name": {
                        "title": "Test group name",
                        "description": "Test group name",
                        "type": "string",
                        "format": "string"
                    },
                    "tests": {
                        "title": "Questions list",
                        "description": "List of test questions",
                        "type": "array",
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/Test"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TestGroup"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}